Basic tests written for local publisher, working on publisher skeleton
